Move beatmap-related classes to beatmap package
