+++ src/com/rian/difficultycalculator/beatmap/DifficultyBeatmap.java
@@ -1,0 +1,4 @@
+package com.rian.difficultycalculator.beatmap;
+
+public class DifficultyBeatmap {
+}
