+++ src/com/rian/difficultycalculator/beatmap/hitobject/HitObject.java
@@ -1,8 +1,8 @@
-package com.rian.difficultycalculator.hitobject;
+package com.rian.difficultycalculator.beatmap.hitobject;
 
 import com.rian.difficultycalculator.math.Vector2;
-import com.rian.difficultycalculator.timings.DifficultyControlPoint;
-import com.rian.difficultycalculator.timings.TimingControlPoint;
+import com.rian.difficultycalculator.beatmap.timings.DifficultyControlPoint;
+import com.rian.difficultycalculator.beatmap.timings.TimingControlPoint;
 
 /**
  * Represents a hit object.
